# Conflicts:
#	src/main/java/br/ufg/inf/pitanga/controller/SistemaController.java
#	src/main/java/br/ufg/inf/pitanga/entidades/Assento.java
#	src/main/java/br/ufg/inf/pitanga/servicos/GeraTabelas.java

+++ src/main/java/br/ufg/inf/pitanga/dao/SalaDao.java
@@ -39,7 +39,7 @@ public class SalaDao implements InterfaceSalaDao {
     }
 
     @Override
-    public Sala buscarPorId(int id) {
+    public Sala buscarPorId(Long id) {
         return manager.find(Sala.class, id);
     }
